backend/requirements.txt
Flask==1.1.1
firebase-admin==3.2.1
google-cloud-datastore==1.10.0
gunicorn==20.0.4
sendgrid==6.1.1
